+++ R/print.parameters_model.R
@@ -66,6 +66,22 @@ print.parameters_model <- function(x,
   # save original input
   orig_x <- x
 
+
+  # check if user supplied digits attributes
+  if (missing(digits)) {
+    digits <- .additional_arguments(x, "digits", digits)
+  }
+  if (missing(ci_digits)) {
+    ci_digits <- .additional_arguments(x, "ci_digits", ci_digits)
+  }
+  if (missing(p_digits)) {
+    p_digits <- .additional_arguments(x, "p_digits", p_digits)
+  }
+  if (missing(footer_digits)) {
+    footer_digits <- .additional_arguments(x, "footer_digits", footer_digits)
+  }
+
+
   # table caption
   table_caption <- .print_caption(x, caption, format = "text")
 
@@ -94,7 +110,6 @@ print.parameters_model <- function(x,
   )
 
   # get attributes
-  res <- attributes(x)$details
   ci_method <- .additional_arguments(x, "ci_method", NULL)
   verbose <- .additional_arguments(x, "verbose", TRUE)
 
@@ -111,14 +126,6 @@ print.parameters_model <- function(x,
     .print_footer_cimethod(ci_method)
   }
 
-
-  ## TODO remove in future update when deprecated
-
-  # print summary for random effects
-  if (!is.null(res)) {
-    cat("\n")
-    .print_random_parameters(res, digits = digits)
-  }
   invisible(orig_x)
 }
 
@@ -146,11 +153,6 @@ print.parameters_brms_meta <- print.parameters_model
                         ci_brackets = TRUE,
                         format = "text",
                         ...) {
-  # check if user supplied digits attributes
-  digits <- .additional_arguments(x, "digits", digits)
-  ci_digits <- .additional_arguments(x, "ci_digits", ci_digits)
-  p_digits <- .additional_arguments(x, "p_digits", p_digits)
-
   format(
     x,
     pretty_names = pretty_names,
@@ -176,7 +178,6 @@ print.parameters_brms_meta <- print.parameters_model
     # get attributes
   sigma <- attributes(x)$sigma
   verbose <- .additional_arguments(x, "verbose", TRUE)
-  digits <- .additional_arguments(x, "footer_digits", digits)
 
   # override defaults. if argument "summary" is called in "model_parameters()",
   # this overrides the defaults...
@@ -218,7 +219,11 @@ print.parameters_brms_meta <- print.parameters_model
   }
 
   if (identical(format, "html") && is.null(caption)) {
-    table_caption <- "Model Summary"
+    if (isTRUE(attributes(x)$is_ggeffects)) {
+      table_caption <- title_attribute
+    } else {
+      table_caption <- "Model Summary"
+    }
   } else if (isTRUE(attributes(x)$ordinal_model)) {
     table_caption <- ""
   } else if (!is.null(title_attribute) && is.null(caption)) {
@@ -271,9 +276,16 @@ print.parameters_stan <- function(x,
   verbose <- .additional_arguments(x, "verbose", TRUE)
 
   # check if user supplied digits attributes
-  if (missing(digits)) digits <- .additional_arguments(x, "digits", 2)
-  if (missing(ci_digits)) ci_digits <- .additional_arguments(x, "ci_digits", 2)
-  if (missing(p_digits)) p_digits <- .additional_arguments(x, "p_digits", 3)
+  # check if user supplied digits attributes
+  if (missing(digits)) {
+    digits <- .additional_arguments(x, "digits", digits)
+  }
+  if (missing(ci_digits)) {
+    ci_digits <- .additional_arguments(x, "ci_digits", ci_digits)
+  }
+  if (missing(p_digits)) {
+    p_digits <- .additional_arguments(x, "p_digits", p_digits)
+  }
 
 
   formatted_table <- format(